OBJ_SceneCacheTransform registers a custom tag reader with HoudiniScene.
This simply translates the tags already existing in the scene file, and does not currently allow for live editing/addition of tags. Also, tags are not maintained when using FlatGeometry mode. In the future, tag addition and FlatGeometry tags may be made available using custom parameters and geometry prim vars.