Restored support for Boost 1.43 (boost::filesystem v2).
Previously we moved to v3 of the boost filesystem library, because v2 is no longer supported in versions >= 1.50. But we still need to build with boost 1.43 to support Arnold on OS X, which links to that version and doesn't hide the symbols, and filesystem v3 wasn't introduced until boost 1.44. This fix reinstates support for 1.43, and adds a todo for when we can finally leave 1.43 behind.