Fixed PTCParticleReader build.
When we switched to specifying includes with -isystem in the CXXFLAGS, the configure check for pointcloud.h started failing, and we stopped building PTCParticleReader. This is fixed by using CheckCXXHeader which uses CXXFLAGS rather than CheckHeader, which doesn't.

Adding PTCParticleReader back into the build revealed another problem - it didn't implement the fairly recent pure virtual positionPrimVarName() method and therefore couldn't be instantiated, so this commit also implements that method.