Added typeName argument to registerRunTimeTyped().
This allows typenames prefixed with the module name to be passed, overriding the typename autogenerated from the class name. There could be an argument for just autogenerating names prefixed with the module name, but that has potential knock-on effects to existing code which is relying on the old names. It's also impossible within python to find the fully qualified path to a nested class, so in that case we'd still want the ability to specify a typename explicitly.

Also removed the deprecated baseClass argument, which was documented as being scheduled for removal in Cortex 6.