Fixed IECoreGL::PointsPrimitive selection.
Previously selection would fail when in Selector::IDRender mode, as demonstrated by the failure of Selection.testPointsPrimitiveSelect(). When in IDRender mode, the Selector installs a special ID shader in the state, and relies on using the override mode to ensure that the shader isn't replaced by others lower down in the hierarchy of the scene. However, the PointsPrimitive replaces the shader so that it can use its own vertex shader to perform the instancing necessary to render as disks and quads, causing the failure. Shader::Setup::ScopedBinding now registers all bound shaders with the current Selector wherever necessary, fixing the problem.