Fixed failing Shader test cases.
These test cases have been broken ever since I rewrote the Shader code for better performance and proper scoped binding/unbinding of parameter values. They were broken because I changed the API rather than because they indicated actual failures. I left them in as a reminder to consider whether or not we needed equivalent functionality to the old APIs they tested, but having gone a good 6 months with the new API I'm pretty happy that it's sufficient as is. So I've patched up the tests so they don't rely on the old APIs.