Fixed IECoreGL::PointsPrimitive colouring.
The PointsPrimitive vertex shader used for instancing didn't have a Cs uniform parameter. This meant that the code in Primitive::render() was generating GL errors attempting to set non-existent uniforms, and also that the fragment shader didn't receive a valid value for the fragmentCs input. The Primitive code had another vulnerability in that it assumed that the location of the Cs parameter was always 0, which was only luckily the case so far.