Fixed OBJ time dependency issues.
Time dependency has a different meaning in OBJs as in SOPs. In OBJs, a node may be time dependant simply because its parent is, since that effects its world matrix. Since Houdini tries to encourage every OBJ to be static, the solution is to believe Houdini when we're told an OBJ is animated. We now only update time dependency of OBJs when Houdini thinks the node is static, but we know otherwise.