Fixed erroneous compute on SceneShape outputObjects.
A Maya bug causes mesh attributes to compute on scene open, even when nothing should be
pulling on them. While dynamic attributes can work around this issue for single meshes,
arrays of meshes suffer from the erroneous compute as either static or dynamic attributes.

We work around the problem by adding a PostLoadCallback if the node was created while the
scene is being read from disk. This callback is used to short circuit the compute which
was triggered on scene open. Since the compute may have actually been necessary, we use
this dummy dependency attribute to dirty the output object attributes immediately following
load. At this point the callback deletes itself, and the newly dirtied output objects will
re-compute if something was actually pulling on them.