Adding setPresets() to Parameter class.
This will allow dynamic interfaces to exist, by updating the shown presets based on value changes in other parameters in the same Parameterised object (most likely Ops).
The idea is to make Gaffer support a user hint that forces a given Parameter to be shown as a drop down list control, and then, use the change-value callback mechanism to update the presets and update the current value if not in the new presets.