Added name parameter to ToHoudiniGeometryConverter.
The blindData name is used for backwards compatibility only, and will be ignored if a name is provided through the parameter.

The Group converter uses blindData childNames on the Groups to name its children in the hierarchy, and in that case, the name parameter can be used to override the top level name only.