Patch for loading multiple ieCoreHoudini plugins.
Houdini loads all plugins on the path, regardless of whether they have been seen before. This means that when installing things in a studio setup, several versions of the plugin may be loaded, for example a user install for testing and the central install. This commit prevents such situations from seg faulting, and instead prints a warning to stderr. Currently the only conflict will be the definition of GU_CortexPrimitive. Since the correct library will still be loaded, I think the warning can be ignored in most situations.