HoudiniScene can use a default time during evaluation.
Added get/setDefaultTime methods, which provide a default cooking time for methods that do not accept time as an argument (e.g. hasObject or childNames). In a HoudiniScene which points at a SOP, it is necessary to use time in these methods. The default time will pass through to children automatically. If left unset, CHgetEvalTime() will be used
for these queries.