Reorganizing parameters on SceneCache OBJs.
Standard OBJ parms are now visible. This required re-jigging things so our new folders are in the same tab layout (switcher) as the originals. Since the original transform parms do not function, the Transform folder is hidden by default.

This causes a strange bug which is fixed by the next commit.