Fixing the unit test in LinkedSceneTest.py ( testLinkBoundTransformMismatch ).
When bbox samples were being accumulated for each call to accumulateBoxSamples(), we were not considering the case that there could be samples already registered and located earlier in time.
The solution adopted was to use the first new sample bbox and use it to enlarge any known bbox that came prior in time.