IECoreMaya::DrawableHolder selection fix (API change)
IECoreMaya::DrawableHolder used to be derived from MPxLocatorNode. This would have been fine and dandy, but for the fact that maya treats locators like flat cards when you're selecting them. Because of this, selection was working very poorly for large objects like environment map spheres. We're now deriving it from MPxSurfaceShape, which gives us more control over the selection process and allows us to implement it correctly.