Fixed IECoreArnold for TBB 4 on OS X.
TBB was including headers which #defined MIN/MAX, which then clashed with the arnold inlines functions of the same name in ai_types.h. The solution is to make sure that ai.h is included before anything else (except boost python, which has its own reasons for wanting to be first).