Added face normals mode to MeshNormalsOp.
This is specified with a new "interpolation" parameter which specifies the interpolation type of the normal primitive variable created. Currently supported values are "Uniform" for face normals and "Vertex" for per-vertex normals as before.

While doing this I experimented with using PolygonIterator.h and PolygonAlgo.h in place of the existing setup. PolygonIterator did make the code much clearer, but added approximately 7% overhead so I rejected it. Using polygonNormal() from PolygonAlgo.h dealt with concave normals and colinear faces correctly, but added ~40% overhead, so I rejected that too because for our current uses we value speed more than correctness.