Preventing double cooking in HoudiniScene::scene( path )
Previously the scene method was not using the default cook time for a portion of its calculations, resulting in double cooking of SOPs which contain a flattened hierarchy. Note that the additional test doesn't actually demonstrate the issue, as it is only apparent in an interactive session, because the viewport forces extra cooks as well. I've verified the cook count is appropriate in interactive mode, and added the test just to be sure no other extra cooks are being caused by the scene method.