Added GIL release for DisplayDriver imageData and imageClose bindings.
This is necessary for some unit tests I'm currently adding to Gaffer for the Display node. The Gaffer display driver node emits signals (with connected Python slots) on the server thread,
so the GIL must be released on the client thread so that those signals can be processed.

Also removed a couple of unnecessary includes.