Now building IECoreRI regardless of compiler version (at IE).
We need to build IECoreRI so we can read/write PTCs in Houdini 13 (which uses a different compiler). Since I removed the compiler version guard, but didn't modify the install paths (we decided not to for now), its possible to build the procedural and display driver with a different compiler, overwriting the versions we actually want. Currently we are just throwing those away (by installing to /tmp when the compiler doesn't match). If we want them in the future, we need to change the install paths.