Initializing defaultObjectPoolCache later in plugin load.
It seems that HoudiniDSOInit is called several times during startup, one of which seems to be in a forked process or something, and we end up with 2 versions of the UT_Cache registered. This results in a broken Cache Manager UI. By moving the cache init to a different function, it only gets called once, and all works as expected.