Fixed compiling on Linux Mint 16
I had to delete a lot of code that was defined but not used, and
probably broke something in the process. I don't know exactly where to
begin testing to make sure I didn't break anything, so any feedback
would be appreciated. Otherwise, Cortex *seems* to compile fine on my
machine.