Added an improved Python wrapping mechanism.
There is significant overhead in acquiring the GIL and entering python looking for virtual method overrides, and this has been showing up in several profiles recently, and also badly affecting thread scaling. Often the overhead was being incurred for classes which were wrapped but for instanced which weren't actually subclassed in python or defining any overrides.

This new mechanism detects whether or not an object is subclassed in Python at construction, and avoids ever entering Python looking for overrides for non-subclassed objects. It also avoids creating reference cycles and the associated garbage collection in this case.

Additionally, the provision of wrapper classes specific to different base classes makes wrapping a large class hierarchy much easier and less error prone. The RunTimeTypedWrapper automatically adds the appropriate virtual method forwarders whereas before every piece of wrapping code had to manually insert the IECOREPYTHON_RUNTIMETYPEDWRAPPERFNS macro. This style of wrapping has been in use in Gaffer for a while and has worked well there.

Currently, none of the existing Cortex bindings are using the new wrapping mechanism - for now it is intended to be used in Gaffer first where the performance benefits will be felt the most.