Further Fixes for Linux Mint 16
- Implemented fix from bentoogood's branch for GCC473 fixes. I return
error 139 with the proposed fix on line 1580 of Sconstruct, so I
replaced it with just printing everything to screen and copying it to
the results.txt file
- SweepAndPruneTest had an unused variable.
- Left the Inverse function of ColorTransform and SpaceTransform
commented out for now. My current theory on why it is not compiling is
that it seems that we're trying to make an object with the
ColorTransform and SpaceTransform classes, and GCC will not allow
classes with virtual functions to be objects. One possible fix is
creating a GenericTransform class and having the functions get derived
from that.