Fixing bug of infinite instantiation of Ops due to knob changes during validate method.
The previous code was forcing changes in the selection knobs for instances of the Op triggered by Nuke during evaluation of nodes. The validation for each instance was modifying the knobs, which would trigger more instantiations.
The solution was to use the Op::firstOp() method which returns the pointer to the main instance, and make sure we call the knob-updating methods only from that instance.