Fixing previous change on the firstReader() implementation.
Some of the functions in the class (hash functions and geometry compute) were using data that was not coming from the firstReader, leading to crashes.

In order to solve the problem and prevent future ones, I've isolated all the member functions that should be shared into a private class that only the firstReader allocates.
Also, all the member functions that are supposed to be used only by the firstReader have an assert() statement to document it.
Lastly, during this process I've realized that some of the data variables that are linked to the knobs were being overwritten during the knob_changed() function, which would create disparity to the non-firstReader instances. So I've created additional ones in the shared data.