Added memory object deduplication for SceneCache.
Added a non-virtual function entryHash() to StreamIndexedIO (todo: we should consider adding to IndexedIO in next major version). This function is supposed to return a unique hash for a child entry.
Using this hash in the cache objects in SceneCache, so repetitive Objects will return very quickly. It fallsback to the previous hash if the IndexedIO is not derived from StreamIndexedIO (see todo above)