Rewrote TestSceneCache to use parallel_reduce.
This fixes an issue where task.errors() was not returning a meaningful result. Now it returns the sum of all errors on all threads. I verified that without the fixes from 030e7c936de091e22564acec4bbffcb222154286 the test still gets the seg fault, and with those changes it passes.