Fixed bug in SceneCache ROP when re-rooting flattened geo with errors.
Previously, if the render SOP had errors, the ROP would continue on trying to write the file, and would return successfully. We know notice that the geo doesn't exist and abort the render. Note that this does leave a partial file on disk, since we aborted after opening the file handle.