Disabling Nuke DeepImageReader test.
Executing a DeepWrite in command line mode was causing a seg fault, preventing the other tests from running. I've submitted a bug to the Foundry (Ticket#2014061310000018) with a simple repro that does not use Cortex. We can re-enable the test if they fix the crash.

I also renamed the test deep exr file, because Nuke's DeepRead didn't like the .dexr extension.