Added function to test LRUCache threading.
This is useful both for testing performance and correctness. The GetterFunction does almost no work, making it easy to measure the raw performance of the cache itself. Ideally we'd have an IECoreTest python module to bind this into, but it's in IECore itself for now.

Also gave PythonLRUCache its own mutex typedef, so it doesn't depend on what should be private implementation details of the base class.