Added private TransformStack class to IECoreRI.
This keeps track of a stack of M44f matrices, optionally with multiple time samples at each level of the stack. It will be used initially in the IECoreRI::RendererImplementation to track transforms before worldBegin(), where the renderman and cortex transforms differ. It may also have uses for emulating a transform stack complete with motion blur support in the IECoreArnold::Renderer class, but until that day I think it's best that it remains private - we have enough bloat as it is.