Added support for multiple cameras to IECoreRI::Renderer.
Each camera is marked with a call to RiCamera(). This meant keeping a track of transforms prior to worldBegin() ourselves, which will pave the way to implementing motion blur for cameras too.

Since we no longer need to defer camera output to worldBegin(), this also means that the hacky "ri:outputNow" parameter to the camera() method is no longer needed, so it has been removed.