Fixed LRUCache crashes where clear() was called concurrently with get().
The problem was that eraseInternal() was assuming that if status==Cached, the cache entry was in the list already, but in fact that is not the case. Another thread may have loaded the entry and set status to Cached, but not have updated the list yet. We now early out of listErase() if the item is not in the list, and unconditionally call listErase() in eraseInternal() regardless of the current status.

Since this bug stems from the same misconception as the hangs in limitCost(), I've also updated the internal documentation in the header to bang home the message that the status and the list items are not necessarily in sync at any given time.