Fixed crash in MeshMergeOp.
This was introduced by the recent despatchTypedData DataPtr->Data * change. When AppendPrimVars went through the `if( m_remove )` code path, it was causing the last reference to `data` to be removed, destroying it. But it would then add data into m_visitedData, making another reference and leading to double deletion.

Moving the `m_visitedData.insert( data )` before the deletion fixes the crash. I'm not sure the visitedData logic was actually right in any case (what if the same data is referenced twice on mesh 1, but should be appended to with two different bits of data from mesh 2) - this commit doesn't do anything with respect to fixing that.