Revert "Changing the test so it clears the cache from time to time."
This reverts commit 668cb72701a18841fbcb52bee3e4b8b2d9822c01.

The testing for clearing is now done in the LRUCache test cases anyway, and now the real problem is fixed, the ComputationCache test is erroring at `BOOST_CHECK_EQUAL( size_t(500), cache.cachedComputations() )`, because the size doesn't hit the cache limit any more due to the repeated clearing. It seemed best to retain the original test rather than remove that check and have two otherwise identical tests in ComputationCache and LRUCache.