Added IECorePython::CastToIntrusivePtr.
This can be used as a return value policy when binding a C++ function returning a raw pointer which must be cast to an intrusive_ptr to allow python to share ownership. Previously it was necessary to manually wrap every single such function to return an intrusive_ptr, and then to bind that.