Defining SceneInterface::hash() as a function that adds the typeId.
Each derived class is now responsible to call the base class implementation as appropriate.
For LinkedScene, that only happens when the location includes the main scene and when the location or the information is purelly derived from the linked scene, the class simply forwards the call to the linked scene.
For SceneCache we always call the base class.