Made ConfigLoader execute files within a directory alphabetically.
Before the order of execution was dependent on the order returned by os.walk(), which can vary from machine to machine and installation to installation. This was responsible for at least two problems in Gaffer going undetected until after release, here's one of them :

https://github.com/ImageEngine/gaffer/commit/e8cc0a236370b34db4e8ac70fd9385e8e2148a70

Note that this does have the potential to expose futher problems in config setups where the config files were relying on whatever order they happened to have been executed in the past - if so those config files will need to be fixed.