Added ExternalProcedural object type.
This derives from VisibleRenderable, allowing a representation of an external procedural to be edited, saved, loaded, copied, hashed, grouped etc. It is suitable for passing through a Gaffer graph, and will form the basis for representation external procedurals in Gaffer. When rendered, it is passed to the renderer as a lightweight Renderer::ExternalProcedural.