Added external procedural support to IECoreRI::Renderer.
ProcDynamicLoad and ProcDelayedReadArchive are supported but there is no support for ProcRunProgram yet. Parameters may be passed to ProcDynamicLoad either from a single ri:data string parameter, or from arbitrary parameters which are serialised automatically into a command line style. This seems a reasonable way of dealing with the "all in one string" limitations of RenderMan procedural parameters, and since Yeti uses a command line style parser, arbitrary parameters should be useable with at least one third party procedural we're interested in.