SceneCache : Make hash() stable across processes where possible
We were using the memory address of the shared data for a scene as the basis of a hash. This was not only unstable across different processes, but also could have produced non-unique hashes where a SceneCache was deleted and a new SceneCache happened to reuse the memory location for its shared data. We now use the filename where it is available and only fall back to the memory address when we have no other choice.

Also made the `sceneCache()` method non-static, since its first argument was a pointer to an instance in any case.