Switched glsl diffuse and specular functions to match 3delight's properly normalized BRDFS
I'm open to putting these under a new name if we want to keep the old versions - if we overhaul how lights are passed, we'll be changing some of this anyway ( like the use of gl_MaxLights ).  But for the moment, this is a fairly nice high quality spec BRDF.  Maybe a little bit too high quality - the Smith G approximation I'm using might be a bit overkill for the viewport, but it's not that complex, and we need something to kill off spec on the back.