Changed IECoreGL::HitRecord::name field from InternedString to GLuint.
The HitRecord is a low level class intended for use with the Selector class which represents names as GLuint, so it makes no sense to use a higher level representation. This will allow the Selector to be used with simple integer names where that is more appropriate. Continuing to use the higher level representation is facilitated by the NameStateComponent nameFromGLName() and glNameFromName() methods.

Binary incompatibilities :

- Changed IECoreGL::HitRecord members