Fixed IECoreGL PointPrimitive test regression.
This fixes the problems introduced by 5ef1dcf77bf48bb8b26c2733b5c7b948e37d1075.

The issue is that IECoreGL::PointsPrimitive overrides shaderSetup() in order to tinker with the vertex shader to perform fast instanced rendering of disks, spheres and quads. Because we do want to allow users to provide their own vertex shaders if they wish, the PointsPrimitive makes the assumption that if a vertex source has been specified explicitly, that the user knows what they're doing and will take care of the instancing too. This interacted badly with the new flatConstant() shader that the Primitive was using to override wireframe shading - because that shader was explicitly specifying a vertex source.

The solution is simple - use the default vertex source in flatConstant(), and grab the uniformCs variable directly in the fragment source. I should have thought of it in the first place.