Added PLATFORM_OVERRIDE arg to IE options.
This allows us to override the platform when picking versions from the IEEnv Registry. We currently use it to comply with the VFX Platform regardless of OS platform. Note that it affects the install location for Cortex and which versions of the dependencies are used to compile Cortex, but we still use the real platform to locate the dependencies once the versions have been determined.