Skipping tests that are known to fail in Maya 2015 and beyond.
This particular problem was introduced in Maya 2015. The bug is in Maya's code for creating a compound array attribute on a referenced node. Instead of creating it with a default array length of 0, it creates it with an array length equal to the number of child plugs. We can observe that bug manifest itself in FloatSplineParameterHandler::doCreate() - we always start out with 3 unwanted array elements. But then FloatSplineParameterHandler::doSetValue() is called and trims away the unwanted elements, leaving us in a good state. That gets us through the test as far as having created the "opHolderReferencer.ma" file. But the bug then manifests itself again when we reload the file - when Maya creates the compound array attribute for the spline (from commands in the reference edits), it starts with 3 array elements (points). Then Maya just sets the values for two of them, and we end up with the third still hanging around with bogus values.