TransformOp now takes a copy of the data before altering it.
This was necessary to avoid unwanted transformation of identical primitive variables. Since the TransformOp takes a list of variables to alter, we want to make sure that we only alter the variables in the list. Prior to this change, a user of the TransformOp would have been un-intentionally altering other primitive variables which point to the same data.