SConstruct : Fixed compiler used for 3delight config checks.
Because we weren't specifying the language to the CheckFunc call, SCons was assuming C and using just "gcc", whereas we may have specified a custom C++ compiler via the CXX build option. Setting the language to CXX for the test ensures we use the correct custom compiler for performing the configuration checks.