Changed SXExecutor parameter interpretation
It turns out that if you give the SXExecutor a V3fVectorData parameter called "v", it complains at you, as it thinks "v" is a texture coordinate and demands a FloatVectorData instead. This caused puzzlement when a lighter tried to use SX to scale a velocity primvar called "v", so I've changed the logic a bit.