Fixed crash when zooming in on coordinate systems in scene shapes
This was happening because IECoreGL coordinate systems were overriding the gl name state. This was confusing SceneShapeInterface::componentBound, which was expecting all the names to be scene paths, and was getting a null SceneInterface object because of that, making it crash. I've fixed/worked around it by making recurseBuildScene() manually convert coordinate systems to CurvesPrimitive, and I've also put a check for a null SceneInterface in componentBound() so it fails a bit more gracefully next time.